Ensure that hierarchy options are passed on to provider function
This commit changes how providers are constituted so that they now
contain the actual hierarchy entry. The options from the entry is then
passed on, possibly augmented with a "path" entry, in calls to the
actual provider function.

A small sample of how to add a provider function dynamically is included
in the lookup/lookup_test.go file.

Closes #25