Container and httpserver improvements, doc updates
This adds the ability to pass variables into the http query
using '?var=varname:value' query parameters.

The container build now creates a top level directory /hiera
which is suitable for bind-mounting a sidecar or data directory
from the host.

The documentation is updated with the behaviour of the rest server.