Enable rendering of explanation output
This commit turns the node objects used by the explainer into pcore
objects so that they can be serialized properly. The serialization is
then used when a combination of --explain and --render-as is given on
the command line. This enables explanation output to be rendered as
JSON or YAML which in turn is a first step towards a distributed Hiera
with full explanation support.