Add ability to dynamically load and call external RESTful lookup plugin.
This commit adds a special plugin loader that kicks in when a lookup
function is missing. The loader will make an attempt to load a RESTful
plugin with the same name as the function from a "plugin" directory.

The location of the plugin can be altered by configuring the "plugindir"
and "pluginfile" attributes in the hiera.conf file.

The commit removes the current terraform and azure-vault plugins from
the hiera code-base. They are instead made available as plugins in
repositories of their own. See:

https://github.com/lyraproj/hiera_terraform
https://github.com/lyraproj/hiera_azure

Closes #35