Handle special case of rendring an empty map in JSON
An empty map is reflected into the type `map[interface{}]interface{}`
by pcore. This is correct. There's no way of deciding key or value
type of an empty map. JSON is however only capable of marshaling
maps with string keys. This commit ensures that the corner case that
occurs for empty maps are handled correctly.

Closes #45