Ensure correct handling of plugin lifecycle
This commit ensures that loaded plugins have the same lifecycle as the
session that loads them. It does this by attaching the actual plugin
registry to the session together with a loader that caches loaded
plugins. Plugins are loaded in the session are then reused until the
session itself terminates and all loaded plugins are killed.

Closes #55