Add ModuleLookupKey function
This commit adds a new top-level lookup_key function called
ModuleLookupKey. This function expects a module path in a session
option using the global provider.ModulePath as key. The function then
expects a lookup key to consist of several segments split by a double
colon. The fist segment in such a key is considered a module name. If
this name maps to a directory found in the module path and that
directory contains a hiera.yaml file, then a lookup will be done using
that configuration.