Ensure that leading slash is removed when using loader path
The plugin loader builds a hierarchy of loaders that matches the
path in the file system for where the plugins are located. In
windows, this means that the first part of the hierarchy is named
something similar to "C:" and hence results in a path that starts
with "/C:". That path cannot be used without first stripping of
the leading slash. This commit ensures that this happens on
windows.