Give special meaning to --facts flag for Puppet compatibility
In Puppet, there's the concept of "facts" stored in the scope used for
lookups, both at the top level and in a special hash under the "facts"
key. Puppet fills the "facts" hash (and the top scope) by executing the
facter program which collects information from the machine where the
lookup is performed.

Go hiera doesn't really have "facts", but it does have a special option
named `--facts` which used to be an alias for `--vars`. This commit
changes this slightly so that variables obtained using the `--facts`
option ends up both in the top scope and in a a "facts" hash accessible
from the top scope, thus mimicking the Puppet behavior.

Closes #79